refactor: move persistTerminalExecution to its own module
Review on #9463 flagged that executionRegistry imported a value from
AgentRunLifecycle while AgentRunLifecycle imports ExecutionRegistry as a
type. Type imports erase, so there is no runtime cycle today, but the
value import closes the graph: moving the call to a synchronous path
would then fail with a TDZ error. Hosting the helper in its own module
means neither caller depends on the other.

Also switches the agentIdentifier guard from truthiness to an undefined
check so an empty agent name cannot silently drop the field from warn data.
